# openapi.yaml
...
openapi: 3.1.0
info:
  title: photo-portfolio-store
  version: 0.1.0
  description: |
    Photographer portfolio + print storefront API.

    Bounded contexts: auth, events, media, search, catalog, commerce,
    compliance, admin. Each operation declares its context via tag.
servers:
  - url: http://localhost:4000
    description: local
...
security: []
paths:
...
  /v1/events/{eventId}/uploads/init:
    parameters:
      - in: path
        name: eventId
        required: true
        schema: { $ref: '#/components/schemas/Uuid' }
    post:
      operationId: initUpload
      tags: [media]
      summary: Initialize a resumable multipart upload
      x-status: stub
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadInitRequest'
      responses:
        '201':
          description: Upload session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSession'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/uploads/{sessionId}/chunk/{partNumber}:
    parameters:
      - in: path
        name: sessionId
        required: true
        schema: { $ref: '#/components/schemas/Uuid' }
      - in: path
        name: partNumber
        required: true
        schema: { type: integer, minimum: 1, maximum: 10000 }
    post:
      operationId: uploadChunk
      tags: [media]
      summary: Upload a single multipart chunk
      x-status: stub
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: Chunk accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadChunkAck'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/uploads/{sessionId}/complete:
    parameters:
      - in: path
        name: sessionId
        required: true
        schema: { $ref: '#/components/schemas/Uuid' }
    post:
      operationId: completeUpload
      tags: [media]
      summary: Finalize a multipart upload
      x-status: stub
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Upload completed; photo row created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/uploads/{sessionId}:
    parameters:
      - in: path
        name: sessionId
        required: true
        schema: { $ref: '#/components/schemas/Uuid' }
    delete:
      operationId: abortUpload
      tags: [media]
      summary: Abort an in-flight upload
      x-status: stub
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Upload aborted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict with current resource state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unprocessable:
      description: Validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            trace_id:
              type: string
...
    UploadSessionStatus:
      type: string
      enum: [in_progress, completed, aborted, expired]
    PhotoStatus:
      type: string
      enum: [processing, ready, hidden, failed, takedown]
    DerivativeKind:
      type: string
      enum: [thumb, preview, web, full]
    UploadInitRequest:
      type: object
      required: [originalFilename, contentType, totalBytes, chunkSizeBytes]
      properties:
        originalFilename: { type: string }
        contentType: { type: string }
        totalBytes:
          type: integer
          format: int64
          minimum: 1
        chunkSizeBytes:
          type: integer
          minimum: 1
    UploadSession:
      type: object
      required:
        - id
        - eventId
        - photographerUserId
        - originalFilename
        - contentType
        - totalBytes
        - chunksReceived
        - chunkSizeBytes
        - status
        - expiresAt
        - createdAt
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        eventId: { $ref: '#/components/schemas/Uuid' }
        photographerUserId: { $ref: '#/components/schemas/Uuid' }
        originalFilename: { type: string }
        contentType: { type: string }
        totalBytes: { type: integer, format: int64 }
        chunksReceived: { type: integer }
        chunkSizeBytes: { type: integer }
        status: { $ref: '#/components/schemas/UploadSessionStatus' }
        expiresAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    UploadChunkAck:
      type: object
      required: [sessionId, partNumber, etag, chunksReceived]
      properties:
        sessionId: { $ref: '#/components/schemas/Uuid' }
        partNumber: { type: integer }
        etag: { type: string }
        chunksReceived: { type: integer }
    PhotoDerivative:
      type: object
      required: [id, photoId, kind, objectKey, bytes, width, height, watermarked, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        photoId: { $ref: '#/components/schemas/Uuid' }
        kind: { $ref: '#/components/schemas/DerivativeKind' }
        objectKey: { type: string }
        bytes: { type: integer }
        width: { type: integer }
        height: { type: integer }
        watermarked: { type: boolean }
        createdAt: { type: string, format: date-time }
    Photo:
      type: object
      required:
        - id
        - eventId
        - photographerUserId
        - originalBytes
        - contentType
        - status
        - hidden
        - featured
        - createdAt
        - updatedAt
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        eventId: { $ref: '#/components/schemas/Uuid' }
        photographerUserId: { $ref: '#/components/schemas/Uuid' }
        uploadSessionId: { $ref: '#/components/schemas/Uuid' }
        originalBytes: { type: integer, format: int64 }
        contentType: { type: string }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        capturedAt: { type: string, format: date-time, nullable: true }
        status: { $ref: '#/components/schemas/PhotoStatus' }
        hidden: { type: boolean }
        featured: { type: boolean }
        takedownAt: { type: string, format: date-time, nullable: true }
        derivatives:
          type: array
          items: { $ref: '#/components/schemas/PhotoDerivative' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...